+++ razuvaevMusic/WebImageView.m
@@ -22,7 +22,7 @@ NSOperationQueue *queue = nil;
 	const char *cStr = [self UTF8String];
 	unsigned char result[CC_MD5_DIGEST_LENGTH];
 	
-	CC_MD5( cStr, strlen(cStr), result );
+	CC_MD5( cStr, (int)strlen(cStr), result );
 	
 	NSString *output = [[NSString
 						 stringWithFormat:@"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
